UPSTREAM:DP:wifi:ath12k:Fix memory corruption in TID setup and cleanup
Currently, TID is created for all SoCs when a peer is created, but it is
only flushed on the primary peer. This causes the allocated TID on the
non-primary SoC to become corrupted when reassigned to another primary
SoC.

Fixed it by creating the TID only for the primary SoC.

Currently, hal_reo_qdesc setup happens after dma_map, causing incorrect
data to be written directly to DDR during initialization. This leads to
potential data inconsistency due to cache not being flushed before DMA
operations.

Fixing it by moving hal_reo_qdesc setup before dma_map, ensuring that
the data is properly set up and visible to the DMA engine. This avoids
the need for additional cache maintenance operations post-initialization.

Currently, the peer TID pointers are reset inside the REO update loop,
causing cleanup to be skipped if the loop exits early due to pending REO
commands.

This leads to stale pointers in the peer's TID structure, which can
result in memory leaks, invalid accesses, or incorrect REO behavior.

Fixing it by moving the TID reset and REO cache clear outside the loop
to ensure consistent cleanup. Also introduce a temporary variable to
avoid overwriting the original TID pointer.

Fixes: Ie1f91000ab9b ("wifi: ath12k: Modularization of peer object
Part-2")

Change-Id: I30f55088486e529909e4c0e65b9e6b6418f17030
Signed-off-by: Girish Kumar M P <gmp@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>