QSDK: DP: wifi: ath12k: Remove unused rx tlv words
removed the unused rx tlv fields to reduce the cache misses
which will improve the KPI performance.

in support of fse meta data use in rx path there some additinal words
added in the rx_tlv which is casuing the KPI decrement.
As part fixing this we have identify there is two Dwords which can remove from regular
tlv subscription, with this change we got ~800Mbps improvment in TCP UL.

Patch-dependency: None

Change-Id: I24524c0049d62f108713f1c45280caf1d1143464
Signed-off-by: KARTHIK KUMAR T <kartt@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>