UPSTREAM:PROTOCOL:wifi:mac80211: Fix CSA crash when AP+Monitor is enabled in No Virtual Monitor mode
Fix crash during channel switch in AP+Monitor mode
when No Virtual Monitor is enabled.

When hostapd triggers a channel switch,
it sends a channel switch NL event only for AP links.
Since monitor interfaces are not accounted for in hostapd,
this causes a mismatch between reserved and assigned links.
AP links switch channels while the monitor remains on the old channel,
leading to a crash.

To resolve this:

1. Exclude monitor interfaces when counting assigned links,
   as no reserved link context is created for them.
2. Update chanctx_conf and chan_req of the monitor link with
   the new channel context after CSA completion.
3. Add the monitor link to the newly created channel context.

Change-Id: Ie79034dadbd044088aa019f6fa7ccb923a7792db
Signed-off-by: Aniruddha Mishra <anirmish@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>