UPSTREAM:PROTOCOL:mac80211: fix dropping frame from unmatched freq of same radio
Sometimes mac80211 receives a frame from other bandwidth freq which will
not be same as the configured channel center freq, hence mac80211 drops
that frame.

Frames of same hardwares are needed to be processed should not be
dropped so, compare the frequencies by radio idx using
ieee80211_get_radio_idx_by_freq and give it to upper stack.

Change-Id: I14c6b6a0250f1eed3bd402e7ed293e1722a9e2ca
Signed-off-by: Karthikeyan Kathirvel <kathirve@qti.qualcomm.com>
Signed-off-by: Naveen Sivasamy <naves@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>