QSDK: DP: wifi: ath12k: Skip skb with zero length during monitor process
Currently, while enabling monitor mode, driver receives an RX_HDR TLV
without any content and without a corresponding BUFFER ADDR TLV. This
results in an skb being created with fragments (nr_frags > 0), but
with skb->len set to zero, as no payload is present.

Previously, the code did not check for this case and would continue
to process the skb even when its length was zero. This could lead
to issues such as passing empty frames to the networking stack or
triggering warnings and bugs in subsequent code paths that assume
skb->len > 0 when fragments are present.

To address this, add a check to skip skbs that have a non-zero
fragment count but a total length of zero. This prevents any
operations on such invalid skbs and avoids potential out-of-bounds
accesses or erroneous behavior.

Fixes : I96784d025c9
(Upstream: DP: Add non linear skb support for monitor mode)

Change-Id: I73f6fb348ba560625ecf0b9cd26c3096ae75837d
Signed-off-by: P Praneesh <ppranees@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>