UPSTREAM:PROTOCOL: wifi: cfg80211: add band_mask to regdomain and apply it in custom band handling
Introduce a new `band_mask` field in `struct ieee80211_regdomain` to allow
selective application of regulatory rules to specific bands. Update
`reg_process_self_managed_hint()` to honor this mask when invoking
`handle_band_custom()`, ensuring that only the intended bands are
processed. Also, ensure that band_mask is correctly set by validating it
against the regulatory rules being updated. For each band whose regulatory
rule has changed, the corresponding bit in the band_mask must be set.

The band_mask is interpreted as follows:
If the band_mask is zero then all the bands are processed.
If the band_mask is non-zero then all the bands for which the bits are set
are processed.

One use case for this is in AFC (Automated Frequency Coordination), where
only 6 GHz band channels need to be updated when an AFC power event is
received.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Path-dependency: None

Change-Id: If6b0313be915958ece7eee961d571f7059a9316b
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>