UPSTREAM:PROTOCOL: wifi: ath12k: Add support for 6 GHz client type
Currently, the default client type for 6 GHz is set to regular client.
Using regular client may be suboptimal for few regulatory domains
where the Tx powers of regular client is lower than that of a
subordinate client.

Add the 6 GHz client type received from the Firmware to the wiphy
regulatory domain. Also, advertise the support
for regular client in the wiphy regd irrespective of the client type
received from the firmware. This is to ensure that the regular client,
which is the default client type, is always available for use.

Also, parse the TPE IE only for the corresponding client type.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch
subsys/0340-UPSTREAM-PROTOCOL-cfg80211-mac80211-Store-6-GHz-powe.patch

Change-Id: I671e07522869fe5e6ba17fd0097814b582f30544
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>