UPSTREAM:PROTOCOL:wifi:mac80211: Enhance TTLM processing to support offload mode in client
When the expected duration expires in TTLM, the current implementation
enables the links back by setting the active and dormant links in host
using ieee80211_vif_set_links() API. The firmware is not notified about
the changes in the state of links. This results in firmware and host not
being in sync about the state of the links. This might lead to firmware
crashes. To address this, firmware's capability of managing links
internally by mapping the TIDs to link can be utilized. After expected
duration expires, there will be no more TTLM IE observed in the beacons.
When no TTLM IE is found in the beacons after the expected duration,
send the WMI_MLO_AP_VDEV_TID_TO_LINK_MAP_CMDID with default mapping where
all TIDs are mapped to all links. This results in enabling the links back
and host-firmware are ensured to be in sync.

Fixes: None
Patch-work: None
Feature: TTLM
Path-dependency:
0318-0002-UPSTREAM-PROTOCOL-wifi-cfg80211-Add-a-wiphy-flag-to-support-offl.patch

Change-Id: Ie0d60f9e02281fbdfd2818fce4f2284898130a9e
Signed-off-by: Shwetha G K <kshwetha@qti.qualcomm.com>
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>