UPSTREAM:CP: wifi:ath12k: Fix peer cleanup for ML stations
to prevent crash

The current implementation of ath12k_mac_station_remove() skips peer
cleanup for Multi-Link (ML) stations. This causes issues when a link
is removed while the station remains associated. In such cases, the
AP sends a vdev delete command without properly cleaning up the peer,
resulting in a firmware crash due to a non-zero peer count.

To address this, we can enhance the logic to allow peer cleanup for
ML stations during link removal or add.

Change-Id: Ic70abd564d3afd1ea3a98c1f8bf2e03711c7231f
Signed-off-by: Balamurugan Ramar <bramar@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>