UPSTREAM:PROTOCOL wifi: ath12k: Fix get_tx_link
Currently, when obtaining the tx link_id in the case of MLO,
if the STA is not available, the link ID passed or the default vif
link is used.

In the case of ML Reconfiguration Link Removal, if link 0 or
the default link is removed, this leads to a multicast traffic drop.

Fix this issue by taking first available link instead of default vif
link.

Patch-dependency: ML Reconfiguration Link Removal

Patch-work: None

Change-Id: I8aa4a6abeff9e25b8d1bdaa45d1e4547dda7b78e
Signed-off-by: Yuvarani V <yuvarani@qti.qualcomm.com>
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>