UPSTREAM:CP:wifi:mac80211: Fix Fix CSA crash when AP+Monitor is enabled in No Virtual Monitor
Fix crash during channel switch in AP+Monitor mode
when No Virtual Monitor is enabled.

When hostapd triggers a channel switch,
it sends a channel switch NL event only for AP links.
Since monitor interfaces are not accounted for in hostapd,
this causes a mismatch between reserved and assigned links.
AP links switch channels while the monitor remains on the old channel,
leading to a crash.

To resolve this:
1. Create reservation for monitor link as well. This is done in the
   context of __ieee80211_channel_switch for one of the AP links
2. When reservations are used in __ieee80211_csa_finalize for AP links,
   use the reservation done for monitor as well.
3. Both while reserving and using the reservation, we iterate through
   the monitor list and handle it for the monitor having the same band
   as that of AP under processing.

Change-Id: Ia91f1b2b10d51b6d86e8fa07e409e2202b5ea10a
Signed-off-by: Aniruddha Mishra <anirmish@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>