UPSTREAM:CP: wifi: mac80211: fix incorrect status_data_idr assignment for zero status_data(idr's id)
Currently, status_data_idr is not set when status_data is zero, due to the
use of a truthy conditional check. This causes valid status tracking
information to be skipped, leading to potential issues in frame
acknowledgment handling and TX status reporting.

Fix this by explicitly checking for non-negative status_data values before
assigning status_data_idr. This ensures that zero is treated as a valid
(idr's id)value and status tracking behaves as expected.

Change-Id: I0ace92a25b9a7b91fc39f9f20dbb44144e37717e
Signed-off-by: Maharaja Kennadyrajan <mkenna@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>