UPSTREAM:CP: wifi: ath12k: Update peer puncture pattern correctly
Currently, the puncture pattern of the peer is sent to the firmware
as part of the peer assoc WMI command.

Consider a case, the AP is operating in a 5 GHz channel 36 with a BW
of 80 MHz. Now, a channel switch is initiated to same 5 GHz channel 36,
BW 80 MHz, but with a puncture pattern of 0x8. This puncture pattern is
sent correctly to the firmware as part of the WMI comamnd
'WMI_PDEV_MULTIPLE_VDEV_RESTART_REQUEST_CMDID'.
However, the pattern is not updated correctly for the connected peers
after MVR. Thus, host sends the older pattern (0x0 in this case) to the
FW. This causes the FW to send packets in 80 MHz full BW, even though 20
MHz is punctured.

There are two issues here:
(1) The target channel's puncture pattern is not updated in the arvif
    structure after MVR. So, any new STA that tries to connect to the
    AP will use the old pattern.
(2) 'WMI_PEER_CHAN_WIDTH_SWITCH_CMDID' is not sent to the FW
    after MVR for the connected peers. So the FW is not informed about
    any changes to the peer's channel width or puncture pattern.
    Peer channel width switch WMI command will be sent only if the
    'IEEE80211_RC_BW_CHANGED' flag is set. Currently, this flag is
    not set due to an early return due to pending CSA check after MVR.

To fix this,
(1) Update the target channel's puncture pattern in the arvif
    structure after MVR.
(2) Set the changed flag 'IEEE80211_RC_BW_CHANGED', if there is a
    change in the puncture pattern after a successful MVR irrespective
    of any other condition. This will ensure that the
    'WMI_PEER_CHAN_WIDTH_SWITCH_CMDID' is sent to the FW for all the
    connected peers.

Fixes: Ia3eca0b5d02 ("CP: wifi: ath12k: PKG upgrade changes set 10")
Patch-work: None
Tag: PEER_HANDLING
Patch-dependency:
ath12k/0279-003-ath12k-fix-RU-puncturing-pattern-for-peers-after-CSA.patch

Change-Id: Ie6a50c4259bbdb4ae56b223f63f7e2369144de3a
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>