UPSTREAM:DP:wifi:ath12k: Add sanity check for fragment length
In backpressure scenarios, the hardware logs packets with a length
less than 16 bytes (which includes the 802.3 header and padding).
This causes an incorrect calculation when subtracting 16 from the
packet length, resulting in a large invalid value and potentially
leading to a crash.

Adding a sanity check to ensure that the packet length is greater
than 16.

Patch-dependency:
0495-04-Upstream-DP-wifi-ath12k-Support-Ethernet-packet-capt.patch

Change-Id: I0b8dbddff41d67ebec770e4dddd4a01ffe414cc5
Signed-off-by: Sushant Butta <sbutta@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>