Upstream:DP:wifi:ath12k: Zero MPDU metadata region in skb->data after allocation
Currently, driver allocates a new skb for each rx header fragment in the
monitor path. The first 64 bytes of skb->data are getting utilizied for
storing MPDU metadata. However, after allocation, the contents of
skb->data are not initialized and may contain random or stale data
from memory. Failing to initialize the metadata region can result in the
use of uninitialized or junk values when processing MPDU metadata.
This can cause incorrect behavior, data corruption, or hard-to-debug
issues in the monitor datapath, especially if the metadata is later
read or interpreted before being properly set.

Explicitly zero the first 64 bytes of skb->data (the metadata region)
immediately after allocating a new skb, both for the initial fragment
and for any additional fragments. Use memset to clear this region,
ensuring that all metadata fields start in a known state and
preventing the use of uninitialized data.

Also, when a drop TLV is detected during monitor RX parsing,
the implementation frees the buffer address of the corresponding
status buffer but returns HAL_RX_MON_STATUS_PPDU_DONE. As a result,
the PPDU is passed to the restitching logic, which expects at least
two fragments in the skb. However, the current MPDU for this PPDU
contains only one valid fragment, leading to warning messages during
processing.

Fix this by returning HAL_RX_MON_STATUS_DROP_TLV instead of
HAL_RX_MON_STATUS_PPDU_DONE after freeing the status buffer on a
drop TLV. This ensures that the remaining status buffers and buffer
address TLVs for the PPDU are properly freed, and the PPDU is not
passed to the restitching logic, preventing invalid fragment access
and associated warnings.

Patch-dependency:
0495-04-Upstream-DP-wifi-ath12k-Support-Ethernet-packet-capt.patch

Fixes : I4966ec2a21c (UPSTREAM:DP:wifi:ath12k: add support for ethernet
packet capture in monitor mode)

Change-Id: I4d30cb1a730bf6659dd9db42deba135fabe382e4
Signed-off-by: P Praneesh <ppranees@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>