UPSTREAM:CP: wifi:ath12k: Ensure proper locking when fetching arvif from ahvif list
The arvif is retrieved from the ahvif list, which is protected by RCU.
However, access to this list also requires holding the wiphy lock, as
it is used during link addition and deletion. This change ensures the wiphy
lock is acquired when accessing the ahvif list to fetch the correct arvif.

Change-Id: I6495bead7fa2bd2bb364853789b249a42d7aaed4
Signed-off-by: Gautham Kumar Senthilkumaran <gauthamk@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>