QSDK: CP: wifi: ath12k: Update TID for UL service class
For UL only service class, TID value is not updated in the NL
message sent to the user to display the service class.
Also the service class ID provided in the display command is not used
directly to fetch the service class attributes, resulting in printing
the wrong service class.

Update the TID also in the NL message for a UL only service classes.
Use the service class ID provided by the user to fetch the service class
attributes and print them.

Change-Id: Ia4a7514b3b3dfab9bf9b129be98e18fc7272d22a
Signed-off-by: Vivek Chettri <vchettri@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>