QSDK:DP:wifi:ath12k:Update splitphy ds bank id for all links during reassignment
Add support to update splitphy ds bank id for all the links when a link
gets a new bank id. This is because mere assignment without updating the
refcount for that bank profile results in corruption of the entire table
with respect to number of users when a double decrement happens during
vdev delete for every link.
To avoid this, ensure whenever a bank_id for a link is assigned to
another link which happens when a VP profile is shared, the refcount
for the bank is updated accordingly.

Fixes: 016c6c2fe1e2 ("QSDK: DP: wifi: Package upgrade changes for datapath")

Patch dependency:
0528-QSDK-CP-wifi-ath12k-Add-locks-while-using-ppe_vp_pro.patch
Patch-work: NA

Change-Id: Ia8083950d14f827c60810fc845c42a1c4cdce8fd
Signed-off-by: Ramanathan Choodamani <rchoodam@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>