UPSTREAM:PROTOCOL:wifi:ath12k: Fix channel context count
for MLO scenario

During channel update, ath12k_mac_update_rx_channel checks for the
number of channel context for the particular radio(ar). Currently,
the count is returned as 3 for 3-link MLO by ath12k_mac_num_chanctxs_iter.
The iterator is checking for local chanctx instead of radio level chanctx.
Since MLD has 3 links, chanctx count is returned as 3. Due to this,
rx_channel is being set to NULL which leads to AWGN info validation
failure.

In order to fix this, ensure that the iterator is updating the channel
context count only for that particular radio. This will ensure that the
ath12k_mac_num_chanctxs(ar) is for MLO case as well. This helps in updating
of rx_channel properly for that particular radio(ar). When next AWGN event
is triggered, it passes as now, rx_channel is not NULL.

Patch-dependency:
      0230-0001-ath12k-Add-support-to-handle-AWGN-interference-for-6.patch
Feature: AWGN

Change-Id: Ie0cac1ff25d867f6904c599c47af008cec8ad086
Signed-off-by: Nidhi Jain <nidhjain@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>