UPSTREAM:PROTOCOL: wifi:nl80211: Fix disabling of FILS/UBPR
Currently, when FILS max interval or UBPR interval is set to zero to
disable FILS or UBPR, the feature is not disabled. This is because
the parsing logic fails to indicate that there is an update for the
feature.

Fix the issue by setting the update flag to true, when FILS
max interval or UBPR interval is zero.

Change-Id: I4dc61131338572476461ac5dc4d401f88a2a6072
Signed-off-by: Suraj P Kizhakkethil <suraj.kizhakkethil@oss.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>