UPSTREAM:PROTOCOL:wifi:mac80211: Fix dtim and beacon interval display issue
For WDS AP, the station dump currently displays the DTIM and beacon
interval as 0. This issue occurs because the values are fetched from
deflink->conf instead of link->conf. Fix this by using link->conf to
retrieve the correct DTIM and beacon interval values.

Fixes: I798d85deda7 (UPSTREAM:DP wifi:cfg80211: correct the patches number)

Patch-depdendency: 0314-wifi-mac80211-extend-support-to-fill-link-level-sinf.patch

Patch-work: None

Change-Id: Iac80e1e3e2c6ce6ad910539df6dcf5b3035b84b8
Signed-off-by: Suraj P Kizhakkethil <suraj.kizhakkethil@oss.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>