UPSTREAM:DP:wifi:ath12k: Init ahsta to NULL and add NULL check for ahsta
In the API ath12k_dp_primary_peer_migrate_setup(), pointer variable ahsta
is not initialized. This can result in random memory access in failure
case of when peer is not found, and ahsta points to random address.

In access to fix this, initialize ahsta to NULL and also add NULL check
for ahsta in failure case handling.

Patch-dependency:
0339-0008-wifi-ath12k-Handle-dp-tid-setup-and-cleanup-for-UMAC.patch

Change-Id: I8c826b34864ac5f9ee06068acebda21da820009f
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>