UPSTREAM:PROTOCOL: wifi: ath12k: Fix WMI SET TPC for client SP
Currently, for a client which is connected to a Standard Power (SP)
AP, ath12k driver sends WMI SET TPC command with PSD powers from the
TPE IE in the AP's beacon. THis power is limited by the regulatory
power for each subchannel. The regulatory power is calculated as
the SP AP power for the subchannel - 6 dB (mandated power difference
between AP and Client).

This logic fails for cases when one or more subchannels are punctured
and the punctured channel is not SP capable. In this case, the
regulatory powers is calculated as 0 and hence the PSD power is
sent as MIN(TPC_IE power, 0) which results in 0 power for the
punctured channel.

To fix this, use the primary channel for calculating the regulatory
power instead of the each subchannel. Also, use Default Client SP
power for lookup instead of AP SP power as SP AP powers may not
be available in the regulatory database always.

Also, add debugs to log the TPC computation.

Fixes: Ia7b7b513fd9 ("UPSTREAM: PROTOCOL: wifi: ath12k: Add client SP mode support for PSD and EIRP-based reg TPC")
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
ath12k/0440-f-UPSTREAM-PROTOCOL-wifi-ath12k-Add-client-SP-mode-sup.patch

Change-Id: I9f59ae11e89ad8d7c14980093b40f03fac20fdc6
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>