UPSTREAM:DP:wifi:ath12k: Free DP QoS context in case of FW recovery
The DP peer QoS context is allocated when the first MSDUQ is created for
a QoS flow and it is deleted as part of the peer delete.

With FW recovery, the DP peer object is deleted in a separate path
so the QoS context if already allocated should also be deleted.

Delete the QoS context if alloacted, as part of the peer delete in case of FW
recovery.

Fixes:
I4ada24b6105 UPSTREAM: wifi: ath12k: Add support for Wifi QoS

Patch-dependency:
0415-0003-UPSTREAM-DP-wifi-ath12k-Add-support-for-QoS-Datapath.patch

Change-Id: Id04b8f400e4be22f85602f9f5ad8ed2e6d0a6c53
Signed-off-by: Vivek Chettri <vchettri@qti.qualcomm.com>
Signed-off-by: Pavankumar Nandeshwar <pnandesh@qti.qualcomm.com>
Signed-off-by: Hariharan Ramanathan <hramanat@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>