UPSTREAM:PROTOCOL: mac80211: Fix TPE parsing logic for punctured AP
Existing logic used to parse TPE IE from AP beacons and probe responses
does not work appropriately when the AP is operating in a
punctured channel.

Consider a case where the AP is operating in EHT 80 MHz BW and one of
the channel is punctured. The following diagram depicts the channel
allocation and the TPE IE received from the AP.

EHT 320:    |  |  |  |
Punctured:  X
HE 20:         |
Chan Idx:   0  1  2  3

TPE N:         |
TPE K:      |     |  |

TPE IE:    |---|   |---|---|---|
          N  1    K  0   2   3

When rearranging the TPE IE based on the channel index, the subchannel
powers need to be ordered correctly to put each subchannel in its right
place for the EHT 80 MHz channel advertised by the AP. But the current
logic of calculation of the TPE IE offset does not handle these cases
correctly and ends up rearranging the TPE IE incorrectly as:

Expected TPE IE:  |---|  |---|   |---|---|
                 K  0   N  1    K  2   3

Actual TPE IE:    |---|   |---|---|---|
                 N  1    K  0   2   3

This happens as we fail to calculate the offset of HE portion of the TPE
IE within the EHT 80 MHz. This is calculated by downgrading the AP's
chandef from the operating bandwidth until we reach the HE channel width
and calculating the number of subchannels that are lost when we
downgrade the channel width. This logic assumes that one half of the
subchannels are lost when we downgrade one level. This assumption
breaks when the AP is operating in a punctured channel as the number of
subchannels lost when downgrading the channel width may not always be
half of the subchannels and depends on the position of the primary
channel and the puncture pattern.

Fix this by calculating the offset simple as the position of the
HE chandef within the EHT chandef.

HE chandef can be calculated by downgrading the AP's chandef to
meet the HE channel width.

The difference between the start frequencies of the EHT chandef and the
HE chandef divided by 20 MHz gives the offset of the HE portion
within the EHT channel.

Offset = start_freq(HE) - start_freq(EHT) / 20 MHz

EHT 320:    |  |  |  |
HE 20:         |

With the suggested change, the offset for the above example will be
computed as 1. Thus the TPE IE will be rearranged correctly as:

Expected TPE IE:  |---|  |---|   |---|---|
                 K  0   N  1    K  2   3

This logic works correctly for both punctured and non-punctured
channels.

Change-Id: I83bf3cb2d6651dc778f3a40ac29b4defb62cc39b
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>