QSDK:PROTOCOL: Send usenol pdev param during dfs_block_radar_events
The AP interface goes down when a DFS radar is simulated on a 5 GHz AP
operating on DFS channel 52, configured with the dfs_block_radar_events
setting during its CAC period.

After CAC completion, when the vdev start command is sent to the
firmware, it appears to treat this as a DFS violation and responds
with a vdev start response event containing an invalid status code.
This causes the AP interface to go down. Ideally, the firmware should
disable the NOL when dfs_block_radar_events is configured on the AP.

To fix this issue, send WMI_PDEV_PARAM_USE_NOL to the firmware with a
value of 0 (indicating that the firmware should disable its NOL)
when the dfs_block_radar_event command is executed.

Tag: DFS
Patch-dependency: 0515-QSDK-PROTOCOL-wifi-ath12k-Introduce-missing-dfs_bloc.patch

Change-Id: Ifc31792ccd80a17c580cf4be61012f707c0c6b73
Signed-off-by: Hariharan Basuthkar <hbasuthk@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>