QSDK:PROTOCOL: wifi: ath12k: Fix Low and High frequency ranges
Currently, during mac registration, the low and high freq ranges for
a radio is limited by both HW range and the regulatory domain range.
The regulatory domain range here corresponds to the default country
code sent by firmware. Currently, the default country code is set to
"US". Thus, accordingly, any channels outside the US regulatory
domain are disabled during mac registration.

This makes those channels unavailable for use even if the user
configures a different country code later that allows those channels.
This is because any channels that are disabled during mac registration
are remembered and remain disabled later even if the regulatory domain
changes to a country that supports this channels. The intention here
is that those channels that are disabled during mac
registration are not supported by the hardware and thus should not
be enabled later. However, in this case, the channels that are
disabled during mac registration are actually supported by the
hardware but are outside the default regulatory domain.

To fix this, limit the radio's low and high freq ranges only by
the HW range and not by the regulatory domain range during mac
registration. This will ensure that all channels supported by the
hardware are available for use when the user configures a different
country code later.

Fixes: Id55eeffab16 ("ath12k: fix hw chan advertisement logic")
Tag: REGULATORY
Patch-work: None
Patch-dependency:
ath12k/0073-wifi-ath12k-Fix-frequency-range-in-phy-info.patch

Change-Id: I11d36db75fb26495fdc99d3e7ad3508796f6015e
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>