UPSTREAM:CP: wifi: ath12k: Enable pdev suspend and resume command
The pdev suspend command helps reduce power consumption. When the last
vdev on a particular pdev is deleted, the host sends a
WMI_PDEV_SUSPEND_CMDID to the firmware. In response, the firmware sends
back a WMI_PDEV_SUSPEND_EVENTID to the host.

Similarly, when the first vdev is created and the corresponding pdev is
in suspend mode, the host sends a WMI_PDEV_RESUME_CMDID to the firmware.
The firmware then responds with a WMI_PDEV_RESUME_EVENTID.

This commit address the missing hunks of pdev suspend and resume.

Change-Id: Ic95fd0c0f39db7a3484c0967c0dff81d1fae6c93
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>