UPSTREAM:DP:wifi:ath12k: Protect the dp->peers list with dp_lock in recovery path
Protect the dp->peers list with dp_lock in recovery path while accessing it
to delete the link peers.

To achieve this, save the mac address and vdev_id in separate per ar tables
in the first loop with dp_lock held, And in the second loop call the
dp_peer_delete without holding the dp_lock.

The above split is necessary, as calling the dp_peer_delete in the first loop
itself will result in deadlock.

patch-dependency:
0344-0002-QSDK-CP-wifi-ath12k-Add-support-for-Mode-1-recovery-.patch

Change-Id: I87ab0f823dccca8d5d7f9299f001f999399c11e9
Signed-off-by: Pavankumar Nandeshwar <pnandesh@qti.qualcomm.com>