UPSTREAM:PROTOCOL:wifi: ath12k: Fix NULL pointer dereference
The ar pointer is being accessed without verifying if it is properly
initialized, which could lead to a NULL pointer dereference.

Fix by initializing the pointer to NULL and adding a NULL check to
ensure ar is only accessed when it is valid.

Patch-dependency: 0381-0003-UPSTREAM-PROTOCOL-wifi-ath12k-Add-TTLM-
offload.patch

Patch-work: None

Fixes: I68dc41c75226(UPSTREAM:PROTOCOL: ath12k: Add TTLM
negotiation and WMI support for peer mapping)

Change-Id: Ib0b11ca043756771faf80cc4407ee51e5e341a52
Signed-off-by: Dhanavandhana Kannan <dhanavan@qti.qualcomm.com>