UPSTREAM:CP:wifi:ath12k:Validate max ML peer id from firmware
Max ML peer id supported by firmware will be sent through mac setup
complete event. This will be the maximum ml peer id, a host can allocate
to peer incrementally. The bitmap maintained in host is 1024 and if
firmware supports post that, peer list will be corrupted as the ml peer
is used as index for peer list in dp_hw.

Max supported ML peer id is derived from number of radios times of
maximum number of ML station, used the same for declaring the bitmap in
host. In firmware supports beyond the value, add a condition to limit to
host maximum

Patch-work:none
Patch-dependency:
	0286-QSDK-CP-wifi-ath12k-Implement-handshake-for-ml-peer-.patch

Change-Id: I44cdff90c5192d1871a30d8aa486c3167c2322a3
Signed-off-by: Ramya Gnanasekar <rgnanase@qti.qualcomm.com>