UPSTREAM:PROTOCOL: wifi: nl80211: Fix update ssid during update AP operation
Currently, the AP update operation is not correctly updating new SSID.
This issue arises because the NL80211_CMD_UPDATE_AP command does not
process its NL80211_ATTR_SSID attribute. Resolve this issue by ensuring
that the SSID is updated if it is provided with the
NL80211_CMD_UPDATE_AP command.

Change-Id: I6d24aded1791d30f16e5f2667518971523152635
Signed-off-by: Nikhil Gurram <nikhgurr@qti.qualcomm.com>