QSDK:CP: wifi:ath12k: WAR to avoid the warn-on during channel switch
In the current design, the monitor interface is added to the assigned
links in mac80211 along with the AP channel context. As a result, both
monitor and AP interfaces share the same list of channel contexts for
assigning or unassigning them from the driver.

During a channel switch command, the driver initiates CAC. If another
channel switch command is issued during this time, a warn-on is printed
in the ath12k driver. This warning occurs when the link is not properly
removed and a channel context change is attempted. The root cause is
that the monitor interface is added alongside the AP interface, so the
number of assigned channel contexts always greater than zero.

During the second channel switch, mac80211 assumes radar detection and
tries to change the channel context without properly removing the link.
To prevent this, a workaround is added to block channel context changes
when radar detection has not actually occurred.

The radar-enabled flag should only be set when CAC is actively ongoing.
This is determined by comparing the total CAC time and the radar
detection flag before setting the radar-enabled flag in the channel
context.

Change-Id: I405f5ca1d70b897e127070a4bc43fe2aba6a991f
Signed-off-by: Gautham Kumar Senthilkumaran <gauthamk@qti.qualcomm.com>