QSDK:PLATFORM:wifi:ath12k: Remove minidump reference segments
Adding changes for the below,

- Remove the minidump reference segments from the kernel minidump driver
when the minidump collection is done.

This is done to fix the usercopy issue when minidump read happens.
ath12k_link_vif structure has one deflink entry which is allocated and
freed in the mac layer. Hence this does not get tracked in the
ath12k driver kzalloc and kmalloc wrapper.

The ath12k_link_vif structure gets added to kernel minidump driver,
but it was not freed since this deflink is not tracked from the
ath_debug module.

Hence adding the changes to remove the reference segments after
its collected at the time of crash.

Change-Id: I36b11c6e5d51afdaa53aaa86faf49d66395c14bb
Signed-off-by: Raghavendran Madhavan <raghmadh@qti.qualcomm.com>