UPSTREAM:CP:wifi:ath12k: Avoid WARN_ON() for bridge sta when not needed
In 4-chip devices, during a 3-link STA connection, a bridge vdev may
be created depending on the order of link creation. Once all three
links are successfully created, the bridge peer will not be created
as it becomes unnecessary.

However, during client association, if the bridge vdev exists but
the bridge peer has not been created, a newly introduced WARN_ON()
triggers a call trace. This is misleading in scenarios where
the bridge peer is legitimately not required.

Prevent this warning by skipping the WARN_ON() when the bridge peer
is not expected.

Depends-On:
    I9cd7115c710c0e3d74376825990bfc3826407f0f
    Id4f0cf2bc9fa517671ecd5bb26985126dc227233

Change-Id: Ib0c667cac5aa62947c559b13c7d86c7347aa6fec
Signed-off-by: Nagarajan Maran <nmaran@qti.qualcomm.com>