QSDK: DP:wifi:ath12k: Update macid in fse_metadata
Current code maps the VP WLAN uplink traffic to VP specific rings causing HOLB issue.
With this change we are programming macid in FSE metadata which is used to map the traffic to the respective egress net device rings, thus resolving HOLB issue in the VP uplink Datapath.

Patch-dependency: none

Change-Id: I0dd322e86b42ed913134dbe770c9a6f3bc41f5bc
Signed-off-by: Prateek Sharma <prsha@qti.qualcomm.com>