UPSTREAM:CP: wifi: mac80211: Fix usage of uninitialized variable
In ieee80211_reconfig(), sdata is declared and getting used while
calling drv_set_rts_threshold() without initialization. This might lead
to corruption as the argument is checked for NULL, but it is not
necessarily to be NULL in this scenario.

Fix this by initializing the sdata pointer to NULL.

Patch-dependency: 0112-mac80211-add-rts-threshold-support-for-MLO.patch
Patch-work: None

Fixes: 9011c52e0356 ("package upgrade: mac80211 patch set 12")
Change-Id: I62cc66ace5372d6e79589694ab8fd897a3519808
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>