UPSTREAM:DP:wifi:ath12k: Do peer setup for newly added peer
In the current architecture, for link reconfiguration feature, when
a new link is added for a station, then a new peer is created.
For this newly created peer, peer setup is not done due to which when
packets are received on this new link, those are routed to Rx REO
Exception ring and these packets are eventually dropped.

In order to fix this, do peer setup for the newly created peer which
sends WMI command to firmware with REO Destination Indication value
and then the packets are routed to the right REO Desitnation ring.

Fixes: Ib9c5c9dc9b5 (PROTOCOL: wifi: ath12k: PKG upgrade changes ML Reconfig)

Patch-dependency:
0325-0005-wifi-ath12k-Handle-link-sta-delete-for-AP-STA-mode-s.patch

Change-Id: Ibbc3de2e6938ae90db7802138393ae980c0a4905
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>