QSDK:DP:wifi:ath12k: Add src_info (VP num) during primary link migration
Currently, we are passing '0' as src_info during primary link migration.
This leads to VP num as 0 and cause packet drops in PPE.

Fix this by passing the VP num as src_info during the primary link
migration. This helps in updating the AST table with correct VP num.

Patch-Dependency:
	0339-0005-wifi-ath12k-add-support-to-handle-UMAC-migration HTT.patch

Fixes:
	I319730cf68d (UPSTREAM:CP: wifi: ath12k: add support for primary link
		      migration)

Change-Id: I930fc6dd8bc627f7503854bf4cc11e9f80c8a2a7
Signed-off-by: Ganesh Kariganuru Mahabalesh <gkarigan@qti.qualcomm.com>