UPSTREAM:CP:wifi:ath12k: Prevent peer assoc with reassoc flag before station association
When a fixed bit rate is configured, the host sends a peer assoc
command with the reassoc flag set to true via a workqueue for all
connected stations. This workqueue is triggered when a station
transitions from state 0 to 1.

However, if a client is connected but not yet in the associated state,
sending peer assoc prematurely with incomplete ML flags and parameters
can lead to firmware assertions due to inconsistent data.

This change adds a validation to ensure that peer assoc with the reassoc
flag is only sent after the station has transitioned to the associated state,
preventing incorrect parameter communication to the firmware.

Dependency:none
Patch-work:none

Change-Id: I583b8d632513c4a83822dbe5bb5f863737101d40
Signed-off-by: Ramya Gnanasekar <rgnanase@qti.qualcomm.com>