QSDK:CP: wifi:ath12k: [WAR] Fix ROC scan issues
Stop dev during link assign operation if vdev is already started for ROC scan.

Do not proceed hw scan if ROC is in progress. There are some Tx/RX failures if both
the scans co-exists. so do not allow hw scan if remain on channel is in progress.
only proceed to hw scan when is_roc flag is set to false once all the ROC reuqests
completed and no new request received in 3 seconds.

In ROC, vdev is started to do tx/rx. So in MLO ROC, if scan is coming in different
radios one after another vif switch is not possible since vdev is started.
So if incoming scan request freq is belong to different ar than the ar on which vif
is started, then stop the vdev and start again in ar of which roc frequency is belongs to.

Patch-dependency:none
Change-Id: Id5fafcac4800fcdbd668c38128f07d444add8cf0
Signed-off-by: Arunpandi Kannan <arunpand@qti.qualcomm.com>
(cherry picked from commit ec04a81cb491a5b39597302262d48579bf2b0a37)