UPSTREAM:CP:wifi:Fix stack corruption due to bitmap overflow on 32-bit systems
While populating the critical update bitmap in the beacon template,
the vdev ID bitmap is sent to firmware. This bitmap was currently
declared as an unsigned long, which caused stack corruption on
32-bit architectures when attempting to set the 32nd bit.

This issue is fixed by replacing the bitmap with a 64-bit variable
using DECLARE_BITMAP. Since the bitmap is split into upper and lower
32-bit segments for firmware communication, the extraction logic is
updated to handle both 32-bit and 64-bit architectures correctly.

Change-Id: Ifdf36ec2273cd6a9413d3a3f7b29df35f1df6454
Signed-off-by: Ramya Gnanasekar <rgnanase@qti.qualcomm.com>