QSDK:PROTOCOL:wifi:ath12k: Add OFFSET_SIGN for 320MHz
Incorrect channels are added to NOL, when dfs_simulate_radar is executed
on an AP operating on 320 MHz with a offset -130.

The offset is packed only into bits [10:3] without a sign bit which
is sufficient for -128..127 but not for extended negative ranges
needed at 320MHz. Additionally, debugfs code enforced range checks
that unnecessarily block valid 320MHz inputs.

Fix this issue by defining an OFFSET_SIGN as (BIT(13)). When width is
320MHz and the offset is negative, encode offset into OFFSET and
set OFFSET_SIGN. For other cases, encode offset as before.

The fix of issue can be implemented only if the support for bang radar
support in 320 MHz is enabled in the firmware.

Therefore, make the following changes to implement a two way handshake to
enable the bang radar support for 320 MHz:

1. Add a service bit WMI_TLV_SERVICE_BANG_RADAR_320_SUPPORT and a host
   service flag WMI_RSRC_CFG_HOST_SIMULATE_RADAR_320_SUPPORTED.
2. Set WMI_RSRC_CFG_HOST_SIMULATE_RADAR_320_SUPPORTED in the host service
   flags in the WMI_INIT_CMD_ID (sent from Host), if
   WMI_TLV_SERVICE_BANG_RADAR_320_SUPPORT is set by the firmware.

Tag: DFS
Patch-dependency: 0234-02-ath12k-Add-debugfs-extended-support-radar-simulation.patch

Change-Id: I341839ae77d85fa63c6430b502117b5bd8a6365b
Signed-off-by: Hariharan Basuthkar <hbasuthk@qti.qualcomm.com>