UPSTREAM:DP:wifi:ath12k: Handle fragmented packets spanning across multiple buffers
In regular Rx path reap procedure, skb pointer in the Rx software
descriptor is set to NULL first and then descriptor is added to the Rx
descriptor used list.

In case of fragmented packets wherein a packet spans across multiple
buffers, Rx software descriptor is first added to local linked list and
when complete packet is received, then local linked list is added to
the used list.

In case if complete packet is not received or processing is not done due
to the current napi budget limitation, then these SG frames are handled
in the next napi cycle.

With the above implementation in place for fragmented packets in next
napi cycle, setting of skb pointer in Rx software descriptor at the
first place results in NULL pointer access eventually resulting in crash.

In order to fix this, set skb pointer in Rx software descriptor to NULL
only while adding the descriptor to the used list.

Fixes: Ie9068808024 (UPSTREAM: DP: wifi: ath12k: RX path performance improvements)

Patch-dependency:
0465-0004-UPSTREAM-DP-Use-scratchpad-memory-in-rx-path.patch

Change-Id: Idd19da4fdefa9c9bdadbc5724deddacab29e5d12
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>