UPSTREAM:PROTOCOL: wifi:ath12k: Reset to previous country when no regulatory rules are received
When an invalid country code is configured to the firmware through the
WMI_INIT_COUNTRY_CMD, the firmware is unable to find this country in
the regulatory database, and as a result, the regulatory domain data
is cleared and mem-zeroed in the firmware, the number of regulatory
rules is zero.

In ath12k_pull_reg_chan_list_ext_update_ev, if total_num_reg_rules is
zero, the dfs region is set as ATH12K_DFS_REG_UNSET (whose values is
zero) and the function returns -EINVAL. Later, when a 5GHz AP interface is
brought up on a DFS, the dfs region is filled as ATH12K_DFS_REG_UNSET
and sent to the firmware in the vdev start request.

When DFS radar is detected by the halphy, there is an assert triggered
in the DFS radar filtering logic in the halphy, as the DFS region is
set as DFS_REG_UNSET, leading to a crash.

This issue is fixed by resetting to the previous country instead of
leaving the regulatory state undefined, when the firmware provides
zero regulatory rules in the WMI_REG_CHAN_LIST_CC_EXT event.
This is achieved by introducing a new work struct and handler to set
the previous country, and updating the event handling logic to queue this
work when needed.

Tag: REGULATORY
Patch-dependency: None
Patch - [PATCH ath-next] ath12k: Reset to previous country when no regulatory rules are received

Change-Id: Ib2a3e20b4ef4d2acfed7a0eb0cad2df0ebe5286a
Signed-off-by: Hariharan Basuthkar <hbasuthk@qti.qualcomm.com>