UPSTREAM:CP wifi:cfg80211: pass radio index in MU EDCA event
Currently we pass wiphy index to be utilized to identify the radio in
which the MU EDCA values should be updated for.

With the introduction of single wiphy architecture, all radios will
be under one wiphy. As a result, MU EDCA values were being updated
only on the first link, many times being the wrong one.

Add radio index as one of the parameters passed to usersapce.

Additionally, when propagating FW updated MU EDCA values to user
space, ath12k is directly calling cfg80211 function. This renders
mac80211 functions useless as they are not used.

Remove the unused code in mac80211 related to this task

Change-Id: I3e83dfce9b2ced54b99134d49c62afda39c7aa42
Signed-off-by: Muna Sinada <muna.sinada@oss.qualcomm.com>